Introduce FsdbSubManager
Summary:
Higher level library for subscribers that hides a lot of the nitty gritty work of the actual transport of deltas. I'm using patch stream client directly here instead of PubSubManager, I think between this and FsdbSyncManager (for publishers) we can probably slowly retire PubSubManager, since it is a very thin wrapper and has been hard to manage the api sprawl there.

This library by default is built to support multi path subscriptions. The design here was a bit tricky. Initially I wanted to allow a subscriber to attach one callback per path that they subscribed, and have the callback recieve the final thrift (or cow) struct. However this presents a problem in multipath subs, because the subscriber (ex DsfSubscriber) expects the different paths to come in the same chunk and we update our internal state all at once.

To support this type of model, I've opted to always hand the subscriber the fsdb root along with a list of changed paths. Only the subscribed paths of the root struct are populated. The subscriber can then loop through the changed paths list and pull out the respective data. Optionally, there is a concept of SubscriptionKeys built in which essentially is an intenger reprenstation of the path we subscribe to

To build on top of this:
- multi device (and multithreaded) version of this for DsfSubscriber and maybe SwitchAgent
- raw thrift version instead of cow
- simplified single path version that doesn't require exposing fsdb root

Differential Revision:
D57745642

Privacy Context Container: L1125642

fbshipit-source-id: 0d3f79d4148f35feb07edf8e7defe74d97b00708