Adding handlers for switch reachability change notif and data structures
Summary:
In the DNX ASIC, we get a RTP table change notification when DSF node reachability
changes. RTP table and related terms are specific to DNX and hence generalizing this
to be a fabric reachability change notification and adding support for the top and
bottom half handling in SaiSwitch. Top half would enqueue the request and return,
whereas the bottom half would do the needed queries into SAI layer to get the actual
details of fabric reachability change and make it available in FSDB.

Reviewed By: jasmeetbagga

Differential Revision: D58413249

fbshipit-source-id: 098a10c1252abd5a5d635adeec17ffd07b6e2878