Export config generation to OSS part 7/N
Summary:
Add new Cmake files for the new utility, as well as a helper script to simplify build + execution of the config generation.

Once this is landed, it should be sufficient to unblock vendors from generating the configs themselves. There are some improvements that can probably be done (e.g. allowing specifying the output directory, more clear errors, option to skip rebuild, etc.), but just this seems like a good landing point. IMO the improvements can be addressed later.

Reviewed By: harshitgulati18

Differential Revision: D58536349

fbshipit-source-id: 7d4b48344c42e83bf6034982d16b6b7641e161f7