replicate mirror session based on number of switch ids
Summary:
As titled, replicate mirror session based on number of asics.

Currently, one mirror session is stored in switch state for all asics but the egress port is local to the asic and will be different across the asics. Hence, replicate the mirror session once per asic so that they can be resolved separately.

Reviewed By: jasmeetbagga

Differential Revision: D58498412

fbshipit-source-id: ac65219c2d9e49d58c9798e850e9eab21087629c