Introduce API to get vlan/port combos needing MAC clear
Summary:
Before a port is removed as part of speed change, need to ensure that all MAC
addresses learnt on VLANs on those ports are removed. Adding a new API which will
give the mapping of each VLAN to a list of ports where MAC addresses needs to be
cleared.

Reviewed By: zechengh09

Differential Revision: D58650109

fbshipit-source-id: d00a14af52a51d0ed78a10b6cd07a8c8e93728cd