Modify expected TxFir setting if zero preemphasis is enabled
Summary:
As titled. Currently we set preemphasis to zero for port serdes, but did not maintain the zero preemphasis setting for next delta processing - this leads to resetting txFir to the desired values from platform mapping.

Fix this by
1. Process change zero preemphasis setting prior to serdes programming.
2. Different handling for leaba and brcm
a. Leaba requires remove and create for zero preemphasis
b. Brcm requires direct setting of zero preemphasis but not accepting create.
3. For desired TxFir setting, if zero preemphasis is set, all Tx setting should be zero.

Reviewed By: jasmeetbagga

Differential Revision: D57993788

fbshipit-source-id: 5d79aef91c95f0692607c165c39e9b5740624dfe