set destination ip api for mirror
Summary:
As titled, api to overwrite mac address of mirror.

For sflow mirror to work, packet has to be sent to eventor port on J3. However, for the eventor to forward it to right destination, L3 lookup has to be performed. Packet being recycled has to be set with eventor port router mac to ensure L3 lookup happens.

Reviewed By: zechengh09

Differential Revision: D58819060

fbshipit-source-id: 8945ae67592014412b5619992ac5dd719cc0b943