Watchdog start test
Summary:
Tests that expected watchdog devices are actually created and device is able to be opened (started).

Note: For this and all stacked watchdog diffs, if we add a requirement that the GETTIMELEFT option be provided it would allow us to more reliably test all of these other features.
Second Note: We will need to update the BSP requirements document to reflect the fact that we don't use system watchdogs anymore (wdts that restart the whole system) but rather a subset of the existing specification applies to other watchdogs such as fan wdts.

Reviewed By: somasun, mikechoifb

Differential Revision:
D58443882

Privacy Context Container: L1125642

fbshipit-source-id: e1bd3bdc79e1854ef68e2aa526ac1ee106f32125