Fix how we copy tx pkt buffer in TxPktEventSyncer
Summary:
IOBufs can be chained so the current impl can throw an overflow since we'll try to read `len` bytes only from the first element in the chain. Instead use another cursor to handle the copy

This was exposed once we switched out of legacy rocket_frame_parser. Removing the workaround

Reviewed By: avalonalex

Differential Revision: D58969181

fbshipit-source-id: 97ce8dbbdb105afa839f044648895fb3bbef4c74