Fix IOBuf usage in IPv[46]::handlePacket
Summary:
These lines are attempting to restrict the cursor to a boundary of the ip pkt size. However how we're doing this is pretty funky and won't work if the IOBuf we have is chained. Fixing by using folly::Cursor's constructor which has an option for the cursor size limit

Recent thrift changes are causing these IOBufs to come as chained from hw switch in multi switch, which causes a buffer overflow here https://fburl.com/everpaste/len7qj2r

Reviewed By: zechengh09

Differential Revision: D59045089

fbshipit-source-id: 7b04d8734abb6904ce3324d6c2ff13c023fc8888