Bypass the transceiver upgrade test when no transceivers are upgradeable
Summary:
Sometime the tcvrToTest list is empty because whatever optics are present in the system, their PN are not mentioned in the list firmwareForUpgradeTest[] in test_qsfp_config_current. So the tcvrToTest list remains empty.

In such case, we need to return true instead of trying to validate the upgrade on all the optics.

Reviewed By: birdsoup

Differential Revision: D59097732

fbshipit-source-id: 619632441349a79d51155bb48a3f7289779ebd20