Cancel pending rxpkt enqueue before unregistering sdk callback
Summary:
If the rx packet queue in hw agent is full, then it will block the sdk rx callback handler. If hwagent tries to do warmboot shutdown in this condition, the unregister callback api in sdk fails because it is waiting for a lock held by rx handler which is inturn waiting on the blocking enqueue call.

Fix: When doing hwagent warmboot shut, use cancellation to unblock the bounded queue enqueue call before unregistering callbacks in sdk

Reviewed By: zechengh09

Differential Revision: D59289867

fbshipit-source-id: a8f133a96ff6675b9cacc5404a1c256116236f46