Include the ForceLowPwr bit in our power control bitmask
Summary: In S430337, we've seen some transceivers setting bit 4, which we currently don't check or clear in our transceiver initialization code. We're not yet sure why this bit is getting set (it doesn't appear to be referenced in our code, so it may be a module FW issue). It seems like we should still check this bit when determining whether or not a module is in LP mode, and we should still make sure to clear this bit when clearing LP mode.

Reviewed By: xanabani

Differential Revision:
D59549266

Privacy Context Container: L1125642

fbshipit-source-id: d8dc9c406229205f1543b95211a0660add2a1fca