Optics Config Validation: Add Valid Configs to Qsfp Service Config
Summary:
Adding optional field to qsfp_config to enable passing in a list of valid transceiver configs, based on `transceiver_validation.thrift`. Making field optional to enable onboarding on a per-platform basis and to acknowledge that the field isn't required in a valid qsfp_config file.

In later diffs, not having the field `validTransceiverConfigurations` in qsfp_config enforces that the validator service will not run.

Reviewed By: srikrishnagopu

Differential Revision: D58829108

fbshipit-source-id: 2d9312d4d10d2049019ecec0e26b75c690150ecd