Optics Config Validation: Add QSFP HW Test
Summary:
After discussions with peers, we decided that having a QSFP HW Test in addition to a link test would be beneficial so we could do a faster test just based on vendor name and vendor part number. This replicated the functionality and logging of the link test but within a QSFP test.

This test may fail since there were a few transceivers in the test pool that weren't deployed in the fleet – but this test will always be marked as Known Bad in configerator (D59191363) so it never blocks package qualification. Will wait until that diff is merged in and this test is green to land.

Reviewed By: srikrishnagopu, xanabani

Differential Revision: D59099616

fbshipit-source-id: 73316c7f07eda9dc15bd396a8240dc862a3a70c9