Optics Config Validation: Invalidate Transceiver Config on Invalid Eeprom Checksums
Summary:
We discovered a lot of variation in Vendor Name and Vendor Part Number attributes when doing a full production scrape of Darwin / Fuji / Elbert. A large majority of these variations were associated with transceivers that had invalid Eeprom checksums.

We are now considering all transceivers with invalid Eeprom checksums as "not validated", given that we can't rely on any data read from these devices.

Reviewed By: xanabani

Differential Revision: D59343196

fbshipit-source-id: dbc194b7423663cc132feb4191335b3d13ffeab7