Optics Config Validation: Add Logic to Ingest Modified Qsfp Config Field
Summary: Creates `TransceiverValidator` class, which is responsible for initializing the data structures for performing transceiver validation and performing actual validation as well. Decided to make a separate class for better code organization and maintainability in case more validation steps are needed in the future. This class will also be responsible for performing logging when an invalid transceiver is detected.

Reviewed By: birdsoup, xanabani

Differential Revision: D58946472

fbshipit-source-id: d1394454bfc75b9339c0227f128b0a1f40f1bae9