Optics Config Validation: Add Link Test Setup Support
Summary:
The implementation of the transceiver config validator tool requires some form of access to agent in order to get information on the programmed port speed. As such, I added validation logs to EmptyLinkTest to verify transceiver validation.

Because EmptyLinkTest is used in other locations, this integration simply just adds validation logs to the output of the test – it doesn't affect the failure / success status.

The main use case for this feature is to determine whether or not all ports on a link test device are validated with the source of truth config list. If a user notices the message `Transceivers [xxx] have invalid configs.`, they would then go to the qsfp_service logs to figure out which exact attributes the tests were failing on.

Added function to common file to more easily incorporate summary logs into other LinkTests if required later.

Reviewed By: birdsoup, xanabani

Differential Revision: D59044550

fbshipit-source-id: d6547ffe7ce2f1adb07a285b9e50b68be8e98436