50G support in SffModule
Summary:
This diff stack is towards adding support for configuring 50G on 100G optics for W400.
Hardware wise - for 50G we need to configure the optic in exactly the same way as we do for 100G mode. Only the programming functions in our software need to be enhanced to also expect 50G speeds. This diff adds that

Reviewed By: xanabani

Differential Revision: D59562644

fbshipit-source-id: ae483162e420a086a6c706abb857bcb76549fee6