Fix portNameToMediaLanes and portNameToHostLanes mapping for 50G speed
Summary: Since we don't configure the optics any differently between 50G and 100G speeds, there is no way from hardware to find out which speed is configured (which is what we do for Cmis modules). Thus we need to rely on SW cache and hard code the lanes for 50G vs 100G speeds.

Reviewed By: xanabani

Differential Revision: D59562708

fbshipit-source-id: 7736e6b33dabd0f484c55072dfa8e48e55f19b17