Qualify namespace in fboss/agent/hw/sai/store/tests/TamStoreTest.cpp +1
Summary:
Putting `using namespace` in a header file is a [_bad_ anti-pattern](https://stackoverflow.com/questions/1452721/whats-the-problem-with-using-namespace-std) and we have multiple linters and compiler errors to prevent it. But to get global coverage of `-Wheader-hygiene`, we need to draw down existing instances.

This diff removes qualifies the namespaces of entities in one more files. This will later allow us to remove a `using namespace` from a header file. Making these changes at scale is challenging. If you'd prefer to make the change in a different way, please commandeer, edit, and land.

Reviewed By: palmje

Differential Revision: D59682908

fbshipit-source-id: d4ff4433a4531a6a86c73d85bbe9f4f22ac180a6