Fix OpticsTest.verifyLatches link test for multi-port optics
Summary: This test disables the ports and verifies latches. However, when ports are disabled the corresponding transceivers go back to discovered state when not all ports of a transceiver are disabled. This is the case that happens for multi-port optics. When the transceivers go back to discovered state, the portNameToHostLanes and portNameToMediaLanes maps get cleared and causes this test to fail.

Reviewed By: srikrishnagopu

Differential Revision: D59721468

fbshipit-source-id: e5d027a2a5ce6a0557b6f5146ff4588269336606