Fix package-fboss.py
Summary:
package-fboss.py is a script that copies binary files and shared libs from the
OSS build outputs into a tar to be copied over and run on a switch.

This change fixes an issue where FBOSS binaries would be omitted from the
tarball because they are now located under `{scratch}/build/fboss` instead of
`{scratch}/installed/fboss`.

Reviewed By: harshitgulati18

Differential Revision: D59827651

fbshipit-source-id: da0f74f170615a8d419f7d34c806859782d86c00