Optics Config Validation: Move Transceiver Config Jsons from Stdout to Temp File
Summary:
tl;dr No longer using stdout for printing out stringified JSONs for each non-validated transceiver config.

To log non-validated transceiver configurations to Scuba for Link Tests, the Link Test binary calls a Thrift API in the qsfp_service which, after validating a transceiver, if detecting it is non-validated, creates a folly dynamic object, populates its fields, stringifies it and sends it back to the caller. The caller (LinkTest binary) then `printf()` this command to stdout. We then would parse all of stdout, which includes all logs from a test run, using a `find_jsons_in_str()` function and only create Scuba samples for the relevant JSONs.

harshitgulati18 noticed an issue with this code on `meru800bia`, in which the parser function was getting stuck somewhere. While it's likely this was just a JSON parser bug, we decided that it would be inefficient to parse all of stdout for a few transceiver configurations each time. As such, we are now using a separate file to store these transceiver configs, which are then pulled into a local machine and passed to the parser function. The functionality should be equivalent to previous functionality.

Reviewed By: xanabani

Differential Revision: D59809554

fbshipit-source-id: 988955f423680db35c3b4d79f9864f84f0e81482