Create flag for handling hop limit exceed packets in RX path
Summary:
On J3 devices, we create rx flood by creating the trap that keeps forwarding ttl=0 packets. Once these packets are processed in CPU, they will go through the path of handling ttl exceed (sending icmp hop exceed packets).

This is not the intention of rx benchmark - the purpose was to measure the handling of dataplane punted UDP packets. Therefore, creating a flag to indicate whether we process and send ttl exceed packets - the flag is disabled only for J3 devices.

Reviewed By: jasmeetbagga

Differential Revision: D59867180

fbshipit-source-id: 1aa2d920a9aa8ef5bcfb30c4a31a7d152bc4f14c