Optimize initial sync simple
Summary:
Currently our initial sync serving goes like:
1. Loop through all subscriptions needing initial sync individually
2. skip over subscriptions for which publisher is not ready
3. Visit the subscription path
4. encode
5. serve

Steps 3 and especially 4 are expensive and are done for every single subscription. We can optimize a good amount if we change initial sync to be closer to how we handle normal subscription serve:
1. Traverse subscription path store for subscriptions (only ones needing initial sync)
2. In tandem traverse our state tree
3. Whenever we reach the end of a subscription path, encode state
4. serve all subscriptions at that path

With this we'll both only visit each path once. More importantly, we can cahce the encode once and just copy the binary to each subscription's stream. Will do the caching in next diff

Differential Revision: D58839259

fbshipit-source-id: d66aa3758812f52795a5d101f45a952f14b81c0b