Move subscription data into a dedicated module
Summary: I want to move the locking of subscriptions into SubscriptionManager instead of subscribable storage. This will give us control to relax the lock for adding new pending subscribers. Adding this module as a thread-unsafe store for all the subscriptions. SubscriptionManager will then be in charge of the locking policy

Differential Revision: D58832844

fbshipit-source-id: 74d69979196f08a7c886e990ccf246dcc00f3160