Cache oper state across subscriptions on initial sync
Summary:
Now we will be serving every subscription at a given path in the same visit call. So we have the opportunity to encode a single time, cache it and just copy to each subscription.

Adapting OperDeltaUnitCache to also support just a raw encoding and reusing for path and patch

Differential Revision: D58926449

fbshipit-source-id: 941af373d3cc2de51e86906771fda49dcbd114cb