Skeleton for DsfSubscription
Summary:
Wanted to pull out the actual subscription, data parsing and state management of a single subscription into a seperate module. Motiviation is to make this more unit testable and make it easier for me to make changes to deltas without worrying about the business logic of the actual dsf state and merge conflicts in DsfSubscriber.

Also might use this as a module for benchmarking subscriptions since it now does not rely on SwSwitch

Differential Revision: D58473262

fbshipit-source-id: 095ecbdd20935b147cda89bb662493942dd4e797