Add VOQ_LATENCY_WATERMARK_BIN feature
Summary:
In DSF, we are making use of the latency watermark bins to figure if any packets are
experiencing high latency, which is higher than expected / normal. Add a new feature
for the same as this will be used in future diffs to decide to collect the latency watermark
stats with SAI_QUEUE_STAT_DELAY_WATERMARK_NS attr.

Reviewed By: jasmeetbagga

Differential Revision: D59792334

fbshipit-source-id: 313aa32ae099eeb6e883facb1ef511f00a59fa6b