Optics Config Validation: Add ValidationStatus Cache
Summary:
Adds a new data structure `nonValidTransceiversCache_` which stores the most recently indicated non-validated transceiverIDs. This cache is updated everytime `validateTransceiverById` is called – this happens in the normal loop of a Transceiver's state machine.

A set was chosen over a map from ID to boolean because the two are semantically equivalent.

Certain functions were forced to lose their const qualifiers because they need to update the validationCache.

Reviewed By: srikrishnagopu

Differential Revision: D59831929

fbshipit-source-id: 716c1fc6f7eabbcdffe62e982ec4524b1d78dff1