Publish firmware versions to ODS
Summary:
Publish firmware versions to ODS under two different keys in parallel:

1. NAME.firmware_version.VERSION : The version and subversion are embedded in the ODS key.

2. NAME.firmware_version : The version and subversion are encoded in a single integer where the version is multiplied by 1000.

Using (1), the count aggregation datatype is straightforward, efficient, and precise. However, this will create new keys / time series for each new firmware version. As the ODS regex operator relies on a cache of ODS keys that may not be updated in realtime, leading to delayed updates in dashboards.

(2) avoids the issues mentioned above, but straightforward aggregations are more difficult. It should be possible to use a histogram view to see how many hosts have a given version at a single point in time, but this won't scale for many different FPGAs/devices and can't show trends over time.

Reviewed By: somasun

Differential Revision: D59099253

fbshipit-source-id: 4157b051d583795355c1d0e5ffe948ae290da407