Add EtherType to ACL entries for J3AI
Summary:
Add EtherType to ACL entries for J3AI. This can be extented to other ASIC platforms if
other ASIC supports/needs this.

Current diff is for HW/agent tests for J3AI. For wedge_agent, this wil be done in configs.

For other tests, it will be done separately.

Reviewed By: shri-khare

Differential Revision:
D60353907

Privacy Context Container: L1125642

fbshipit-source-id: 64e9c3ed077921e51929d100dd3513263c14bbc7