Benchmark Enhancements: Build Qsfp Benchmark Binaries per SDK
Summary:
Following up on a TODO from birdsoup, now adding support to build benchmark binaries per SDK so we can catch regressions in new releases.

`.bzl` and `TARGETS` files were also configured in a way that supports us passing folly benchmark command line args, so we know which benchmarks are in each benchmark file, which is needed for passing in `--bm-regex` and `--bm-list`.

Reviewed By: xanabani

Differential Revision: D59621670

fbshipit-source-id: 517b010046f55981d8ad042ea14e360c7a2f8480