Support respin on the same slot
Summary:
Given a slotPath, we resolve a PmUnitConfig to explore.

If productVersion == 1, use the default PmUnitConfig in `PlatformConfig::pmUnitConfigs`
Otherwise, find a matching productVersion PmUnitConfig in `PlatformConfig::versionedPmUnitConfigs`

=========
Next Steps:
1. Respin support on different slots.
2. Deprecate slotPathToPmUnitName_ in DataStore.

For #1, I think we should forbid re-spin on PmUnits that don't have an IDPROM. Or at least enforce to have an IDPROM in the next versions.

Reviewed By: somasun

Differential Revision:
D60025044

Privacy Context Container: L1125642

fbshipit-source-id: ef01fdaae3162d19ab6a1b0eba5f502f0752c28c