Reset cable len (if set) when port goes down
Summary:
Cable lenghts are fairly expensive to collect - 50ms
per port and can change only on cabling changes. So we will
optimize this to only collect cable lengths if a port is
up and does not have cable length set. The flip side of
this is that on port down, we should reset the cable
len to be null. Then when port comes up, the next round
of cable len collection will notice that port is down and
collect the stat accordingly.

Reviewed By: zechengh09

Differential Revision: D60491649

fbshipit-source-id: 9cae0e541864f3a104191cc48001611a56c7edc2