Update cable length only on up ports if cable len is not already set
Summary:
Cable length collection is expensive, taking upto 50ms per
port. Cable length can really only change in face of recabling. So
we optimize as follows
- Reset cable len on port down event
- Collect cable len only for up ports that don't have cable len set.

The reason for resetting on port down event and not on stats collection round
is that stats collection is periodic. So consider a port getting recabled,
if it got recabled and came up within our stats collection interval, we
would not recollect cable len until next warm/cold boot.
Reason for not collecting cable len stat on port Up and doing it in periodic
stat collection is that we may need to try multiple times since when port comes up,
not everything  is synchronized immediately

Reviewed By: zechengh09

Differential Revision: D60491789

fbshipit-source-id: 38028d729b1b79682b677a46260a6db758545e4a