Allow us to move the proto files outside of fbcode
Summary:
X-link: https://github.com/pytorch/executorch/pull/4417

Allow `rust_protobuf_library` to take in a target or a source. This is an effort to move proto files to a standard, shareable location. The decision was made to make users of this macro to explicitly define the export file if they needed.

Reviewed By: ndmitchell, huydhn

Differential Revision: D60187261

fbshipit-source-id: 4a044b07d3096e433d987118f363aba42f54f716