fixing bash
Summary:
bash number starting with 0 are octal, so running the cat command will
fail if it's start with a 0 followed by a 9 bc it will treat it as octal.
Adding #10 will fix this.

Note: This is just a temporary fix because I am working on a new version on
config which won't use shell command.

Differential Revision: D60552833

fbshipit-source-id: 0dc23d6df4c7d89dab2efea3644970f6c18bcea1