Explicitly delete eth.up fb303 stat on port delete instead of deleting from PortStats destructor
Summary:
Swswitch maintains a thread local storage of SwitchStats which in turn contains a PortStats map. The port stats has stats for ethx/y/z.up which is set and cleared using setCounter/clearCounter calls to fb303::ThreadCachedServiceData.

https://github.com/facebook/fboss/blob/main/fboss/agent/PortStats.cpp#L200

The clear is currently called from destructor of PortStats which is not correct since this is TL data and other threads will be still running. Instead explicitly delete the counter when port is deleted. This problem showed up on multiswitch devices because TL stats got destroyed when thrift worker thread exited

Reviewed By: zechengh09

Differential Revision: D60553999

fbshipit-source-id: 11ca8c802b42eec6cec1c36bfa44aebb1aa6afff