Fix broken version readout
Summary:
Fixing these for Whistlers and MP3 because the fix was applied without considering the data types.
```
[root@sdsw001 ~]# cat /run/devmap/fpgas/MERU800BFA_SMB_FPGA0/fpga_ver
0x1
```
FPGA versions are dumped as hex numbers, breaking the "10#, treat as decimal" paradigm.
CPLD versions, however, are output with a leading zero. The question remains if they are actually hex with a bug where someone forgot the x or actually decimal.

For MP3 I'm essentially undoing D60552833 because all of the versions are output in hex with 0x prefix and this breaks them.

We should have this unified - one single format for version output. Also, parsing and interpreting this as a number is prone to breaking and will break again.

**"Read actual, compare with desired, flash if differ" should be our desired outcome.**

Differential Revision: D60652918

fbshipit-source-id: 7e276f2193895be08293e22363e7f1e3ef1bf049