rackmon2: Allow register map to specify if value has sign
Summary:
This allows us to explicitly mark certain registers as
signed or unsigned (default) thus allowing the core to
do any appropriate sign extension on registers and ensure
that they will be decoded as negative instead of large
positive numbers.

A default of unsigned (sign = false) was chosen after
consulting the register maps of the various devices and
confirming that a majority are recommended to be unsigned.

Imported from: D58690175

Reviewed By: alandau

Differential Revision:
D60414281

Privacy Context Container: L1125642

fbshipit-source-id: ae6edcbebbb1dadd8686c7a8f748a40303b69049