rackmon2: Dont use 1s polling for tests
Summary:
This was needed when we did not have the ability
to "tick" the threads and the test code had to
rely on a series of sleeps to test the functionality.

Now those tests have been enhanced, we no longer
need to start the monitor with 1s polling interval
and instead let it pick the default 1min.

This fixes a rare issue where the polling thread
ticks on its own along with our test thus causing
unexpected behavior.

Imported From: D52349709

Reviewed By: tao-ren

Differential Revision: D60415592

fbshipit-source-id: 09e59a5604ae1ae7b8c93d3763a7f5df6154f525