cpp: fixes for unique_ptr and pclose interaction
Summary:
The glibc and/or libstdcpp found in the latest upstream
lf-master refresh does not like using `pclose` to a `unique_ptr`
directly because the return value of the `pclose` is ignored.  Instead,
we can wrap it in a simple lambda so that the return is explicitly
ignored (by casting to `void`).

This is the same changes made upstream for similar code such as:
    https://github.com/openbmc/phosphor-fan-presence/commit/fc2b078ac2969b24c12b22a9d522e5bc73760c8a

Imported from: D54598090

Reviewed By: alandau

Differential Revision:
D60414285

Privacy Context Container: L1125642

fbshipit-source-id: cd9f0329fe9e778c1e133503f6b34dfbd4ac1385