Only look at port drain state when deciding to disable on port loop.
Summary:
Being inactive does not mean creating a loop is ok on
fabric switch port. If a peer VOQ switch is drained, port will
become inactive. However the moment you create a loop on
such a port, the port will start advertising reachability to
itself and create a data plane loop.

The only safe setting to ignore such a loop is if port or
switch itself is drained

Reviewed By: shri-khare

Differential Revision:
D60709211

Privacy Context Container: L1125642

fbshipit-source-id: aad224905ddaf24b4fda9c5747e49b74cbf9bd2f