Fix agent_invariant for RTSW breakage when dlb_crosszone enabled
Summary:
RTSWs have a dlb_crosszone feature where a specific IP network traffic is only DLB'd. This is enabled in only a subset of RTSWs.
This change modifies the IP to easily use the pumpRoCETraffic API to send test traffic and avoid having to parse and use the specific IP in the test.

Reviewed By: zechengh09

Differential Revision: D60723129

fbshipit-source-id: 48c5ecc675d8976c0ff1838cd88b28377b6ce98d