Sai Benchmark Enhancements: Modify AgentBenchmarksMain to Handle Benchmark Listing
Summary:
The current `main()` method of `AgentBenchmarksMain.cpp`, which is used to run `sai` and `bcm` benchmarks, always runs initialization code for wedge agent before running any of the actual benchmarks. However, since we also need to support the functionality to just list tests without any wedge agent initialization, this diff adds a conditional to skip initialization when the `--bm_list` flag is passed in.

As such, I added an if statement that checks if the `--bm_list` flag is passed in.
- If `--bm_list` is passed in, we skip certain initialization steps and use `folly::Init` to parse the folly command line arguments and properly load the flag.
- If `--bm_list` is not passed in, we proceed with initialization as it previously was.

I also noticed some random GTest logs when listing tests within a binary very infrequently, so adding a conditional to skip these lines that start with a hyphen in `QsfpBenchmarkFramework`.

Reviewed By: srikrishnagopu

Differential Revision: D60036687

fbshipit-source-id: 6f9cb4611a9efe0e17f65fdf9d2916b5f653275d