Add FBOSS Agent thrift libs to Sandcastle build target
Summary:
We recently ran into an issue where a missing dependency in a thrift lib for agent caused OSS build failures (see D60293182). This wasn't caught by Sandcastle OSS builds because this path was not exercised.

We can add the thrift libs to the automated builds so that we can catch these types of issues in the future.

Reviewed By: srikrishnagopu

Differential Revision: D60596332

fbshipit-source-id: ad479b49ac08afd8265a3a3734c59aa66319da9c