Qsfp Benchmarks: Move getMatchingTcvrIds into Helper
Summary:
In D60213280, we refactored RefreshTransceiver to only use cabledPortPairs indicated in the qsfp_config file that match the relevant MediaInterfaceCode – this guarantees more consistent benchmark test runs.

This diff extracts this logic out into a separate helper function, so we can use it for to-be-added benchmark functions and for reading registers (next diff). This was done by setting `unique_ptr<WedgeManager>` as `shared_ptr<WedgeManager>`, allowing us to pass it in to a helper function that uses it for fetching the relevant data. Decided this would be safe within the context of qsfp benchmark tests.

Reviewed By: xanabani

Differential Revision: D60726358

fbshipit-source-id: 3e433f1b5d94eebf95b39df763e5055eaa01feeb