Transceiver Validation: Add Lab Device Scraping Scripts
Summary:
We initially built the Source of Truth config DBs for each platform using a scrape of all unique transceiver configs currently used in production. This is because we originally assumed that if we validated a transceiver config at one point in time, it would permanently be considered validated.

This, however, is an incorrect assumption because only the transceivers on lab devices are routinely tested with our newest software. As such, we want our Source of Truth DB to be automatically generated from the transceivers plugged into our link test devices.

In order to reach this point, we first need to:
- Replicate the transceiver part numbers in the fleet in our lab device setup – this guarantees we establish a starting point where all transceivers deployed in the fleet are in our lab device setup.
- Determine which transceiver part numbers our lab devices are missing when compared to the fleet (on a per-platform basis).

In line with this, this diff:
- Fetches all lab devices of a specific platform available in the `fboss.qsfp.autotest` pool (`fetch_platform_lab_devices.py`).
- Adds a data fetching script to use `sudo wedge_qsfp_util` on these devices to retrieve vendor and port information (`lab_device_scrape.sh`).

Next diff will add logic for parsing this data, filtering out only cabled port pairs indicated in the qsfp config file, and generating a list of all transceiver part numbers we have for each platform.

Reviewed By: xanabani

Differential Revision: D60878157

fbshipit-source-id: b21cff3f059aff2e4cf014de87ac2cc263337277