Add Print Production Features and Virtual getProductionFeatures Function
Summary:
Define `printProductionFeatures` and virtual `getProductionFeatures` functions to be used in filtering tests by Qsfp Hw production features.

`printProductionFeatures` prints out all the features of a specific test if the `list_production_features` flag is enabled. This information is ingested by the test runner to decide which tests to run based on the features supported by a specific platform config.

`getProductionFeatures` is a virtual function overridden for every actual test that is run, implemented in each test subclass of `HwTest.cpp`. This was defined as a virtual function, not a pure virtual function, to accommodate for the fact that many tests are not tied to a specific production feature, and can choose whether or not to override the default production features list, which is just an empty list.

`IsSkipped()` is required for subclasses that override the original `SetUp()` function to ensure tests don't run when the test binary is run with `--list-tests`.

Reviewed By: xanabani

Differential Revision: D59335281

fbshipit-source-id: a3a487d3dbdc5ba7e2db3aacf152d891be6306fd