Check if subscription to remote endpoint already exists before emplace
Summary:
In case we have multiple instances of DSFNode object pointing to
the same remote endpoint (e.g. MTIA), the current logic would endup
creating a temporary DsfSubscription object, only to delete it just
after. This is not great as it will create unnecessary session churn.
Note that emplace does not guarantee that a new object will not
be created in face of key already existing.

Reviewed By: zechengh09

Differential Revision:
D60981460

Privacy Context Container: L1125642

fbshipit-source-id: e8753c14a8f37b9a74e71c9fe087ec02647661a5