Towards moving updateWithRollbackLogic to DsfSubscription class
Summary:
As noted in previous diff, there is a tangled web
of callbacks b/w DsfSubscription and DSFSubscriber classes for
updating state, performing rollback. This makes DsfSubscription
rather hard to read. So the intention is to move the
logic for work done in context of a DSFSubscription to
the DSFSubscription itself. This is a prerequisite before
we add more complicated policies like async scheduling of updates
to DSFSubscription.
Note that this diff does not kill the original function in DSFSubscriber.
That will be done in stacked diff as there a lots of UTs to migrate
before that code can be retired

Reviewed By: zechengh09

Differential Revision:
D60996689

Privacy Context Container: L1125642

fbshipit-source-id: 5a33f4bbff0ae36d6fdade8d788ee17b216b0ab7