Add test for deleted credits counter
Summary:
Test simply sets up a data plane loop and then checks for
deleted credits. As the port ramps up to line rate it greedily
grabs credits, not all of which can be used up immediately.
In the process some credits are lost and assert against that.
Before exiting for WB, we break the data plane loop so that
it can be setup again post WB

Reviewed By: nivinl

Differential Revision: D61194979

fbshipit-source-id: a62a67395465a1d5ca03322170dc67c9fafff193