Modify AgentEnsembleTest and AgentHwTest teardown to check for warmboot flag
Summary:
Currently, all of Agent Hw test relies on verifyAcrosswarmboots and so they dont worry about the TearDown implementation.

If any test relied on AgentHwTest or AgentEnsembleTest teardown, then they wont be able to support warmboot because it always exits in an non graceful manner.

Modify AgentEnsembleTest and AgentHwTest teardown to check for warmboot flag and pass in option to exit gracefully

Reviewed By: zechengh09

Differential Revision: D60946407

fbshipit-source-id: 2b4064a13ee403e2d8750c23252dc956acbdf394