Cleanup container after completing FBOSS build
Summary:
In order to run the build scripts multiple consecutive times in a row, the
build container needs to be removed (otherwise, an attempt to spawn a new
container with the same name will result in docker complaining that a container
already exists -> premature exit).

This change stops and removes the existing build container. AFAICT killing and
restarting the build container doesn't actually affect incremental builds (as
long as the scratch path is mounted to the same location).

Reviewed By: srikrishnagopu

Differential Revision: D61301820

fbshipit-source-id: 63f25579a5db0efdcc8a5a83e5f1d7fcfa56a8e9