Assert that hwUpdateEventPool is always of size 1
Summary:
Since we want to schedule destruction of DSFSubscription
on the same hwUpdateThreadEvb as was passed to it, kepp
this hwUpdatePool_ size as 1. This should be fine as
hwUpdates are anyways scheduled on a single SwSwitch::update
thread. If we ever need to increase this, then we must take
care to schedule DsfSubscription destruction on the same
evb as was passed to that DSFSubscription object on
construction

Reviewed By: srikrishnagopu

Differential Revision:
D61320000

Privacy Context Container: L1125642

fbshipit-source-id: 7b68fd5e8d6a6248e6611c99887cc25abe39dd5e