Avoid scheduling gratuitous lambdas on hwUpdateEvb
Summary:
If nextDsfUpdate is not empty, then just overwrite
nextDsfUpdate with latest info but don't schedule
another lambda on the hwUpdateThread. The idea here
is that since nextDsfUpdate_ was not empty it implies
we have already scheduled a update on the hwUpdateEvb
and that update will now simply consume the latest
contents.
needsScheduling = nextDsfUpdateWlock->isEmpty();

Reviewed By: peygar

Differential Revision:
D61248826

Privacy Context Container: L1125642

fbshipit-source-id: 8b1c9e8961ca0b8a3452974bb3e386bee0bc006d